Display the metric's suffix in tooltip.
When we mouseovered on a point on a metric, the value was shown but the
suffix wasn't.
It was due to a curly bracket issue, the "tooltip" option wasn't given
as an "option" sub-object, but as a third argument of the Chart
constructor. A curly bracket was closed to early.
This is fixed and the suffix is now displayed.

Related to CachetHQ/Cachet#2848